minor namespace refactor
