+++ src/main/java/net/kelsier/bookshelf/api/model/bookshelf/lookup/AuthorLookup.java
@@ -1,4 +1,4 @@
-package net.kelsier.bookshelf.api.model.bookshelf;
+package net.kelsier.bookshelf.api.model.bookshelf.lookup;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -10,27 +10,25 @@ import net.kelsier.bookshelf.api.model.common.Operator;
 import javax.validation.constraints.NotNull;
 import java.text.MessageFormat;
 
-/*
-id  SERIAL PRIMARY KEY,
-    book INTEGER NOT NULL,
-    format TEXT NOT NULL ,
-    uncompressed_size INTEGER NOT NULL,
-    name TEXT NOT NULL,
+
+/**
+ * Model for database lookup in the author table.
+ *
+ *
  */
-
-public class DataLookup implements ColumnLookup {
+@Schema(name = "AuthorLookup", title = "Author lookup")
+public final class AuthorLookup implements ColumnLookup {
     private static final String DEFAULT_FIELD = "name";
-    private static final String DEFAULT_OPERATOR = "EQ";
-
+    private static final String DEFAULT_OPERATOR = "LIKE";
     @NotNull
     @JsonProperty("field")
-    @OneOf({"format", "name", "uncompressed_size"})
-    @Schema(description = "Search query field", defaultValue = DEFAULT_FIELD)
+    @OneOf({"name"})
+    @Schema(name ="field", description = "Search query field", defaultValue = DEFAULT_FIELD)
     final String field;
 
     @NotNull
     @JsonProperty("operator")
-    @OneOf({"EQ", "NEQ", "GT", "LT", "GTE", "LTE", "LIKE", "UNLIKE"})
+    @OneOf({"EQ", "NEQ", "LIKE", "UNLIKE"})
     @Schema(description = "Search operator", defaultValue = DEFAULT_OPERATOR)
     final Operator operator;
 
@@ -38,9 +36,9 @@ public class DataLookup implements ColumnLookup {
     @JsonProperty("value")
     final String value;
 
-    public DataLookup(@JsonProperty("field") final String field,
-                      @JsonProperty("operator") final Operator operator,
-                      @JsonProperty("value") final String value) {
+    public AuthorLookup(@JsonProperty("field") final String field,
+                        @JsonProperty("operator") final Operator operator,
+                        @JsonProperty("value") final String value) {
         this.field = field;
         this.operator = operator;
         this.value = value;
