+++ src/main/java/net/kelsier/bookshelf/api/resource/bookshelf/Data.java
@@ -8,10 +8,8 @@ import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import net.kelsier.bookshelf.api.model.bookshelf.BookLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.DataLookup;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.DataLookup;
 import net.kelsier.bookshelf.api.model.common.Search;
-import net.kelsier.bookshelf.framework.db.dao.bookshelf.BookDAO;
 import net.kelsier.bookshelf.framework.db.dao.bookshelf.DataDAO;
 import net.kelsier.bookshelf.framework.db.model.bookshelf.BookData;
 import org.jdbi.v3.core.Jdbi;
