+++ src/test/java/net/kelsier/bookshelf/api/model/bookshelf/BookLookupTest.java
@@ -2,8 +2,8 @@ package net.kelsier.bookshelf.api.model.bookshelf;
 
 import io.dropwizard.testing.junit5.DropwizardClientExtension;
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
-import io.dropwizard.validation.OneOf;
 import net.kelsier.bookshelf.MetaBooks;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.BookLookup;
 import net.kelsier.bookshelf.api.model.common.Operator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
