+++ src/test/java/net/kelsier/bookshelf/api/resource/bookshelf/CommentsTest.java
@@ -2,17 +2,14 @@ package net.kelsier.bookshelf.api.resource.bookshelf;
 
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
 import io.dropwizard.testing.junit5.ResourceExtension;
-import net.kelsier.bookshelf.api.model.bookshelf.BookLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.CommentLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.SeriesLookup;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.CommentLookup;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.SeriesLookup;
 import net.kelsier.bookshelf.api.model.common.Operator;
 import net.kelsier.bookshelf.api.model.common.Pagination;
 import net.kelsier.bookshelf.api.model.common.Search;
 import net.kelsier.bookshelf.api.model.common.Sort;
 import net.kelsier.bookshelf.framework.db.dao.bookshelf.CommentDAO;
-import net.kelsier.bookshelf.framework.db.model.bookshelf.Book;
 import net.kelsier.bookshelf.framework.db.model.bookshelf.Comment;
-import net.kelsier.bookshelf.framework.db.model.bookshelf.Series;
 import org.jdbi.v3.core.Jdbi;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
