+++ src/test/java/net/kelsier/bookshelf/api/resource/bookshelf/DataTest.java
@@ -2,18 +2,13 @@ package net.kelsier.bookshelf.api.resource.bookshelf;
 
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
 import io.dropwizard.testing.junit5.ResourceExtension;
-import net.kelsier.bookshelf.api.model.bookshelf.DataLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.DataLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.SeriesLookup;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.DataLookup;
 import net.kelsier.bookshelf.api.model.common.Operator;
 import net.kelsier.bookshelf.api.model.common.Pagination;
 import net.kelsier.bookshelf.api.model.common.Search;
 import net.kelsier.bookshelf.api.model.common.Sort;
-import net.kelsier.bookshelf.framework.db.dao.bookshelf.BookDAO;
-import net.kelsier.bookshelf.framework.db.dao.bookshelf.CommentDAO;
 import net.kelsier.bookshelf.framework.db.dao.bookshelf.DataDAO;
 import net.kelsier.bookshelf.framework.db.model.bookshelf.BookData;
-import net.kelsier.bookshelf.framework.db.model.bookshelf.Comment;
 import org.jdbi.v3.core.Jdbi;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
