+++ src/test/java/net/kelsier/bookshelf/api/resource/bookshelf/TagsTest.java
@@ -2,17 +2,13 @@ package net.kelsier.bookshelf.api.resource.bookshelf;
 
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
 import io.dropwizard.testing.junit5.ResourceExtension;
-import net.kelsier.bookshelf.api.model.bookshelf.LanguageLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.RatingLookup;
-import net.kelsier.bookshelf.api.model.bookshelf.TagLookup;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.LanguageLookup;
+import net.kelsier.bookshelf.api.model.bookshelf.lookup.TagLookup;
 import net.kelsier.bookshelf.api.model.common.Operator;
 import net.kelsier.bookshelf.api.model.common.Pagination;
 import net.kelsier.bookshelf.api.model.common.Search;
 import net.kelsier.bookshelf.api.model.common.Sort;
-import net.kelsier.bookshelf.framework.db.dao.bookshelf.PublisherDAO;
 import net.kelsier.bookshelf.framework.db.dao.bookshelf.TagDAO;
-import net.kelsier.bookshelf.framework.db.model.bookshelf.Publisher;
-import net.kelsier.bookshelf.framework.db.model.bookshelf.Rating;
 import net.kelsier.bookshelf.framework.db.model.bookshelf.Tag;
 import org.jdbi.v3.core.Jdbi;
 import org.junit.jupiter.api.BeforeEach;
